rename custom-element.ts -> createComponent.ts
