+++ packages/react/src/drafts/TabPanels/TabPanels.tsx
@@ -4,7 +4,7 @@ import React from 'react'
 import styled from 'styled-components'
 import {get} from '../../constants'
 import {TabContainerElement} from '@github/tab-container-element'
-import {createComponent} from '../../utils/custom-element'
+import {createComponent} from '../../utils/customElement'
 import sx, {type SxProp} from '../../sx'
 import type {ComponentProps} from '../../utils/types'
 import getGlobalFocusStyles from '../../internal/utils/getGlobalFocusStyles'
